Add building of cpp runtime to dist rule.
Add tasks to checkout version 7.1 of bdwgc from cvs, 
configure it, build it, and locally install it
in a well-known location in the src-cpp tree.

We are not yet actually using bdwgc, but hopefully will be soon.

Build procedures tested on linux-x86.  
Holler if this causes problems on cygwin and I'll come and debug it.