Several changes to the constraint system: (1) The "BruteForce" technique Nate had introduced has been eliminated. Now we check that c |- d by enumerating from d all those constraints entailed by d which do not involve existentially quantified variables, and then checking for each that they are entailed by c. To make this work, I had to fix some bugs in the way these constraints were generated. So the fix is very simple. (2) Got rid of DUMP_EQV global flag. Now eqvs are always dumped. A flag is passed to dump when we need it to produce constraints that do not involve EQVs. (3) A new toString() is defined for a constraint -- it enumerates the EQV variables in the constraint. (4) Added support to check whether an atomic formula is entailed by cycliing through the atomic formulas on the LHS. (5) Added several tests involving existential quantifiers (EQVEntailmentTest.java).
TODO: 
(1) Refactor the constraint solver so that it deals only with the feature-structure constraint system and knows nothing about self variables, this etc. The constraint solver provides the ability to define constraints (including introducing existentially quantified variables), check for their consistency, check if one entails another, support substitution. That's all. Other X10-specific features should be supported on top of this layer  in an X10-specific layer.
(1a) Remove "hidden" flag from EQVs. EQVs should be the same as local variables. 
(2) In this X10 layer, ensure that an explicit environment is kept in the context containing the constraints in effect at the given point in the code. (Nate has some code for this I believe.) Some current test failures need this to be in place. 
(3) Fix outstanding test failures.