WIP in reducing the number of NativeReped classes in XRX.
This commit makes several related changes to the handling 
of NativeRep classes by the code generator.  
  (1) We no longer emit the empty c/h/inc files for a NativeRep class
  (2) #include directives are generated for NativeRep classes using the
      same algorithm/rules as normal classes.
  (3) x10rt17.h no longer includes the header files for NativeRep classes.

As a consequence, the handwritten .h and .cc files backing a NativeRep class 
must follow the name mangling conventions as would have been used by the 
code generator if it had generated files for the class.  This required
renaming a number of files in src-cpp/x10/io  and also the addition of 
empty .h files for the various NativeReped primitive classes (Int.h. etc).