Yet another rework of the RTT initialization story to try to avoid problems with partially initialized and multiply initialized RTT objects.
The approach now is to completely rely on lazy initialization
via calls to getRTT (including Object and primitives in the same protocol). 
Check against typeName != NULL instead of parentsc != -1
to avoid depending on the order in which C++ static initializers
are run. By using a zero (NULL) value as the uninitialized check,
we can eliminate the constructor for RuntimeType and instead rely on
all static data segments being initially zero.  Since many calls to getRTT
happen during static inialization (due to itables), we need a sequence that 
doesn't depend on the constructor for a particular RTT instance being run
before getRTT is called for its class.