Completely functional implementation of place-local storage for C++. Decided to not use PGAS's SVD functionality for now, since that would widen the x10rt API and a non-embarrassing standalone implementation wasn't actually that much work.
Workaround for the case when an uninitialized PlaceLocalHandle is serialized
(can happen when 'this' is captured by a serialized closure within
 a constructor). For now, we'll re-create the uninitialized storage on the
remote node and if it is used, we'll crash and burn.  Arguably the front-end
compiler should reject as incorrect programs where this happens.

Rewrite KMeansDist to use PlaceLocalHandles instead of using statics to 
attempt to get place-local storage (static semantics changing in 2.0). 
KMeansDist now runs to completion.