Thu Sep 03 03:54:47 2009
Implementation of place types.

The key idea is that we add to the context a constrained term (of type
place), representing the current known information about the type of
here.

To make it easier for programmers to write code, we permit the place
argument p of the at expression at(p) e or statement at(p) s to be of
type Ref or of type place. In case it is of type Ref, the at
expression/statement will automatically coerce p to p.location.

p is permitted to be any expression, including a method call. 
In the cases in which p is not representible as a constraint term,
the compiler introduces a new final variable (say q), initialized
with the expression and treats at(p) e as if it were at(q) e. 
The type of q is inferred to be the type of p. Specifically if the
type of p constrains self to be some know place r, q will be constrained
to be equal to r.

At an implementation level, XContext and ClosureDef_c now carry two
additional items, a placeTerm and a placeConstraint. The
typeCheckOverride method of AtExpr_c is modified to establish the
value of the placeTerm and placeConstraint fields in the ClosureDef_c
associated with the body of the AtExpr_c. 

   This is necessary so that type inference for the body of the
   AtExpr_c work correctly. The type of the AtExpr is the return type
   of the closure representing its body. Now a closure body may have
   multiple return statements inside it. The type of the closure is
   taken to be the merge (least upper bound) of the types of the
   expressions returned from within the body of the closure. This
   information is recorded in the associated codeDef(), of type
   ClosureDef.

   The way type inference for a closure/method body works is as
   follows. A resolver is set for the return type, containing a
   reference to the AST for the body at the time of type
   building. This resolver will be forced to execute when the
   returnType is visited during type-checking, and is
   disambiguated. By this point, typeCheckOverride will have been
   called on the parent AST node (atExpr). Unfortunately, this node is
   *a copy* of the node with which the resolver was initialized!
   Hence, we need to ensure that when the resolver runs it has access
   to the constrained placeTerm. We ensure this by storing the
   information in the ClosureDef associated with the AtExpr AST, since
   all copies of the AST refer to the same ClosureDef.

Additionally, we modify the type-checking of X10New_c so that a
constraint clause equivalent to "self.location==here" is added to the
type of the expression.  here, of course, is replaced by the
constrained placeTerm from the environment.

   An important point: We we have to implement "adding the clause
   |self.location==here|" by adding to the type of the X10New_c
   expression the constraints |self.location==placeTerm,
   placeConstraint|. This is the right thing to do since placeTerm may
   be a symbolic variable about which the only pieces of information
   we have are those in placeConstraint.

   Thus if |p| is a final variable, the type-checker can infer that
   the type of |at (p) (new Foo())| is |Foo{self.location==p}|. If
   |p()| is a procedure call with the return type |Place{self==x}|
   then the type of |at (p())(new Foo())| will be inferred as
   |Foo{self.location==_1786, _1786==x}| (for some new symbolic variable
   |_1786|), i.e. |Foo{self.location==x}|.

Handling of this.

Note that in the body of an instance method, it is not necessarily the
case that |this.location==here|. For instance the body of an instance
method may contain an explicit |at| expression or statement which
changes the value of |here|.


To do:

-- Addion of logic for ateach.
-- Addition of logic for distributions.

Sat Sep 05 21:39:46 2009

When a class is pushed into the context, add the constraint |this.location==here|.

Mon Sep 07 08:53:13 2009

Refactored significant chunks of X10MethodInstance_c abstracting large
swathes of detailed code into small, semantically meaningful, reusable
chunks.

See checkCall and instantiate2 in particular.


Used to be that the env was made to contain the binding xi = xtype.self, and xi is the translation of the formal name.

Bug in old code: 
  Hmm.. this is not right. The variable may not be final, hence cannot
  appear in constraints. Well, then it wont matter that we add this.

Mon Sep 07 18:50:45 2009

Ok there is a leakage of constraints into the formal type. The formal
type should always be very small -- because it should just reflect
what is written in the text. But see ll.txt. We are getting formal
types with scores of conjuncts.

Mon Sep 07 20:00:41 2009

Compile time had gone to 14 minutes for XRX! Reduced it to 5 :-) by
recognizing that when formals are being fixed, the constraints must be
copied. (Otherwise the old constraints were being modified in place,
resulting in larger and larger constraints on each successive pass!)

Mon Sep 07 20:01:41 2009

Realized that need to make sure that thisType is updated by
instantiating with the chosen value for this. 

However, need to make sure that the updated this type is not
communicated to the second call to instantiate2 -- the second call
must be given the original thisType so that it can update it with a
chosen UQV. The thisType updated by this call must be the one passed
into checkCall.

Mon Sep 07 20:03:52 2009

Need to figure out how to add here clauses in methods, e.g.
in FastArray:

 def raw(){ dist.onePlace==here }:Rail[T] = raw;


May already be implemented...check.

Tue Sep 08 00:49:19 2009

Fixed here() --- now it simply genrates a local variable, here. 

Let p be the current place as recorded in the context. To check if
receiver r is local (X10TypeSystem_c.isHere(...)), we now add here=p to the antecedent, and
ask if r.location==p.


Another unintended side-effect on a constraint --- need to copy in computeNewSigma.

Need to add code to ensure that the type of a val field of Ref type with an initializer is extended with the clause self.at(this).

isSubtype needs to add here=p to the antecedent. Fixed.

* Look at Reader.x10. Arguments are passed at the type Marshal (an
  interface). Methods are subsequently invoked. We need a way of
  saying that something is local or a value (so that methods can be
  invoked on it). Specifying the clause self.at(here) does not work,
  since the field .location is not defined.

  To Object, added:

    property def loc():Place;
    property def at(p:Place):Boolean;
    property def at(r:Ref):Boolean;

  and these are defined under Ref and Value. 

  Hopefully this means we can now use self.at(p) with interfaces!

* When expanding bodies of property methods, need to make sure it is done recursively?

Fri Sep 11 06:27:26 2009

* Fixed the problem of at clauses on Interfaces by hard-wiring into
  XTypeTranslator that a call |s.at(t)| where |at| has been resolved
  to the method |Object.at(Place)| or |Object.at(Ref)| should be
  considered as |here == tloc| if the type of |s| is a subtype of |Place|,
  and as }s.location == tloc| otherwise, where |tloc| is |t| if the type 
  of |t| is |Place| and |t.location| otherwise.

* Fixed the return type of |here|. Needs to include the clause |self
  == t| where |t| is the current place (obtained from the context).

* Why is null being allowed as a value for a cosntrained type. Need to
  ensure that when we assign null to a variable, we check that the
  null value satisfies the constraint clause associated with the
  variable's type.

* Need to ensure that activity().finishState.pop() will work, where
   activity: Activity{self.at(here)}
   activity.finishState: State{self.at(this)}