Fixed XTENLANG-364. That is, removed syntactic support for type parameters for closures.
Note that several closure tests related to guards are still failing. Will be fixed in future code revision.

Also fixed a bug in ambiguity detection for method calls (where one call resolves to a closure call and the other to a method call). This yielded several ambiguities in existing code (mainly array classes in XRX) which have also been fixed.