Support for ValRailView. No need to add a new class -- simply create a ValRail where the given Rail is used as the backing array. Note: The native C++ implementation needs to be modified so that when such a ValRail is serialized it copies the data from the Rail into the serialization buffer.
The Rail.copyFrom(srcIndex:int, src:ValRail, dstIndex:int, size:int) should be usable for such a ValRail and should now be a purely local operation. On the remote side we must use makeValRailView on the buffer so we can keep a reference to the relvant portion of the buffer instead of copying.