Working version of "OpenMP within MPI" version of HeatTransfer.
Note that this fixes a race condition in the version presented in
the tutorial slides.  In the slides, there were only 3 phases (2 nexts)
in the loop iteration.  You actually need 4 phases (3 nexts) to prevent
diff(z) being reset to zero by the "master task" at the node before the 
"slave tasks" can read diff(z) and notice that they need to do another
iteration.  There are a number of ways to fix this, I choose one that 
was relatively close to the original example.