WIP.
(1) Removed EffectsVisitor (a NodeVisitor) in favor of EffectsComputer (a ContextVisitor).
(2) Moved methods for computing effects into delegates and extensions.
(3) Added appropriate effects to Any and Struct.
(4) Changed implementation of Effect so that it keeps a Safety tag.
(5) Eliminated BottomEffect, use Effects.makeUnsafe() instead.
(6) Added support for an Effect in MethodDecl.

Todo
(1) Convert from x10.compiler.effects annotations to Effect representations. We must have a way of specifying an effect set which is the tightest possible (no mutable location is read or written), and one which is the 
(2) Add "entailment" for effects. An effect set E1 implies E2 if their safety bits are the same, and every effect in E1 is also in E2. (Needs careful definition in the case of atomic effects.) Now we need to ensure that if a method has an explicit effect set specified by the user, then it implies the effect set computed from the bottom up by examining the implementation. 
(2a) We must have a way of syntactically specifying an Effect which is the strongest possible (every mutable location is touched) and one which is the weakest possible (no mutable location is read or written).
(3) Ensure that a method with a declared effect set can only by overridden by a method that has a weaker effect set.
(4) Set up mechanism for lazy computation of effects for a method parallel to the way in which return types for methods are computed (cf TypeCheckFragmentGoal). 
(5) In X10Call_c, ensure that the effect set in the called method's body is transformed into an effect set in the caller's environment.
(6)Change the logic for checking effects -- all we need to do is to ensure that the declared effect for a method entails its inferred effect. Now, in the most common usage, a programmer will simply mark the main method as @Safe@Effect() (assuming that the way to specify the effect set is the weakest possible effect set is to say @Effect()).