Add API function x10.lang.System.heapSize() to return an estimate of the heap being used by the current place.
Provide a decent implementation of this for Java backend and for the
C++ backend when BDWGC is in use.

The C++ non-GC implementation is vacuous because this isn't the kind of question
that can be answered in a purely portable way when the program is just
using malloc to handle memory allocation.