WIP on C++ codegen for XTENLANG-754
C++ backend now supports boxing a struct "up" to an interface
type and doing instanceof/checkcast operations on the boxed interface.

Invocation of interface methods on the "boxed struct" will not work,
yet as we need to augment to codegen for struct itables to generate
an additional itable thunk class where the receiver parameter is 
defined to be an IBox<S> instead of a S&. I'll work on this over the weekend
(don't think it's too hard, but I may not finish it before calling it a night
tonight).

After interface method invocation on boxed structs works, I will merge the 
trunk changes to the newEquals branch and start looking at C++ codegen issues
with Any on the branch.