Merge changes from newEquals branch to trunk.
Since svn mergeinfo on the branch was broken in various ways,
I did the merge by constructing two patches:
  newEquals: -r12334:12349
  newEquals: -r12351:12386
and then applying them and fixing a couple of additional problems.
Since the merge was done via patch instead of svn merge, the
history of the individual commits on the branch,
is only found by looking at the branch, not here in the trunk.  
Sigh.