Now that we are allowing Structs to be implicitly coerced to Interfaces via a boxing operation, when assigning/casting a Struct to an Interface we must generate a class_cast_unchecked<Interface>(Struct) to cause the boxing to happen.
Updated comments in two places we were already doing this to indicate 
why it was necessary and also changed from class_cast to class_cast_unchecked
since the operation is statically type correct.

Fixed codegen for X10Cast_c to also do this when a Struct is being
upcast to an interface that it implements (in the Object upcast case,
we can still emit the code because there is nothing to do).  Again this
can be an class_cast_unchecked because the assignment is statically 
type-correct, but we do need the explicit operation in the generated code.